P10/requirements.txt
asgiref>=3.5.0
Django>=4.0.1
django-cors-headers>=3.11.0
djangorestframework>=3.13.1
djangorestframework-simplejwt>=5.0.0
drf-nested-routers>=0.93.4
PyJWT>=2.3.0
pytz>=2021.3
sqlparse>=0.4.2
tzdata>=2021.5
flake8>=4.0.1
flake8-html>=0.4.1